Rename README.md to README.txt
Conversion From Binary to Decimal

The purpose of this assignment is to help you understand assembly language programming by writing an assembly language program for our simple accumulator machine.

Given an integer containing only 0s and 1s (i.e. a "binary" integer), you are to print its decimal equivalent. To do this, you should use the division operator and the remainder to pick off the "binary" number's digits one at a time from right to left.

Much like in the decimal number system, where the rightmost digit has a positional value of 1, the next digit left has a positional value of 10, then 100, then 1000, and so on, in the binary number system, the rightmost digit has a positional value of 1, the next digit left has a positional value of 2, then 4, then 8, and so on. Thus, the decimal number 234 can be interpreted as 4*1 + 3*10 + 2*100. The decimal equivalent of binary 1101 is 1*1 + 0*2 + 1*4 + 1*8 or 1 + 0 + 4 + 8, or 13.